+++ helion/language/device_print.py
@@ -1,18 +1,14 @@
 from __future__ import annotations
 
-import ast
 import builtins
 from typing import TYPE_CHECKING
 
 from torch.fx import has_side_effect
 
 from .. import exc
-from .._compiler.ast_extension import create
-from .._compiler.ast_extension import expr_from_string
 from . import _decorators
 
 if TYPE_CHECKING:
-    from .._compiler.inductor_lowering import CodegenState
     from .._compiler.type_info import TypeInfo
     from .._compiler.variable_origin import Origin
 
@@ -67,32 +63,6 @@ def _(*args: object, origin: Origin, **kwargs: object) -> TypeInfo:
     return NoType(origin)
 
 
-@_decorators.codegen(device_print, "cute")
-def _(state: CodegenState) -> None:
-    prefix = state.proxy_arg(0)
-    assert isinstance(prefix, str)
-
-    value_asts: list[ast.AST] = []
-    if len(state.proxy_args) > 1:
-        assert len(state.ast_args) > 1
-        ast_varargs = state.ast_args[1]
-        assert isinstance(ast_varargs, (tuple, list)), (
-            f"Expected tuple for varargs, got {type(ast_varargs)}"
-        )
-        value_asts.extend(ast_varargs[0])
-
-    fmt = prefix + " ".join(["{}"] * len(value_asts))
-    call_args: list[ast.AST] = [create(ast.Constant, value=fmt), *value_asts]
-    call_expr = create(
-        ast.Call,
-        func=expr_from_string("cute.printf"),
-        args=call_args,
-        keywords=[],
-    )
-    stmt = create(ast.Expr, value=call_expr)
-    state.add_statement(stmt)
-
-
 @_decorators.ref(device_print)
 def _(prefix: str, *values: object) -> None:
     print(prefix, *values)
@@ -103,4 +73,5 @@ def _(prefix: str, *values: object) -> None:
 # helion/_compiler/<backend>/.  Import them here (at module import time) so the
 # @_decorators.codegen(op, "<backend>") registrations run with the same eager
 # timing as when the bodies lived in this file -- no behavior change.
+import helion._compiler.cute.device_print  # noqa: E402, F401
 import helion._compiler.triton.device_print  # noqa: E402, F401
