+++ helion/language/view_ops.py
@@ -106,13 +106,6 @@ def _(state: CodegenState) -> ast.AST:
     )
 
 
-@_decorators.codegen(subscript, "cute")
-def _(state: CodegenState) -> ast.AST:
-    # CuTe kernels currently execute scalarized pointwise code, so shape-only
-    # indexing used for broadcast setup is a no-op.
-    return state.ast_arg(0)
-
-
 @_decorators.ref(subscript)
 def _(tensor: torch.Tensor, indices: list[object]) -> torch.Tensor:
     # pyrefly: ignore [bad-index]
@@ -155,83 +148,6 @@ def _(tensor: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
     )
 
 
-@_decorators.codegen(split, "cute")
-def _(state: CodegenState) -> list[ast.AST]:
-    from .._compiler.ast_extension import statement_from_string
-    from .._compiler.cute.cute_reshape import _flat_index_from_coords
-    from .._compiler.cute.cute_reshape import _get_node_dim_local_coord
-    from .._compiler.cute.cute_reshape import _get_tile_shape
-    from .._compiler.generate_ast import GenerateAST
-
-    fx_node = state.fx_node
-    assert fx_node is not None
-    input_node = fx_node.args[0]
-    assert isinstance(input_node, torch.fx.Node)
-    input_val = input_node.meta["val"]
-    assert isinstance(input_val, torch.Tensor)
-    output_val = input_val.new_empty(input_val.shape[:-1])
-
-    cg = state.codegen
-    assert isinstance(cg, GenerateAST)
-    df = cg.device_function
-    env = CompileEnvironment.current()
-    config = df.config
-
-    input_shape = _get_tile_shape(input_val, env, config)
-    output_shape = _get_tile_shape(output_val, env, config)
-
-    input_numel = 1
-    for s in input_shape:
-        input_numel *= s
-
-    dtype_str = env.backend.dtype_str(input_val.dtype)
-
-    smem_ptr = df.new_var("split_smem_ptr")
-    smem = df.new_var("split_smem")
-
-    src_coords = [
-        _get_node_dim_local_coord(cg, input_node, input_val, i)
-        for i in range(len(input_shape))
-    ]
-    src_flat = _flat_index_from_coords(src_coords, input_shape)
-
-    if output_shape:
-        output_coords = [
-            _get_node_dim_local_coord(cg, input_node, output_val, i)
-            for i in range(len(output_shape))
-        ]
-        out_flat_base = _flat_index_from_coords(output_coords, output_shape)
-    else:
-        out_flat_base = "cutlass.Int32(0)"
-    lo_flat = f"({out_flat_base}) * cutlass.Int32(2)"
-    hi_flat = f"({out_flat_base}) * cutlass.Int32(2) + cutlass.Int32(1)"
-
-    cg.add_statement(
-        statement_from_string(
-            f"{smem_ptr} = cute.arch.alloc_smem({dtype_str}, {input_numel})"
-        )
-    )
-    cg.add_statement(
-        statement_from_string(
-            f"{smem} = cute.make_tensor({smem_ptr}, ({input_numel},))"
-        )
-    )
-    cg.add_statement(
-        statement_from_string(f"{smem}[{src_flat}] = {{_inp}}", _inp=state.ast_arg(0))
-    )
-    cg.add_statement(statement_from_string("cute.arch.sync_threads()"))
-
-    lo_var = df.new_var("split_lo")
-    hi_var = df.new_var("split_hi")
-    cg.add_statement(statement_from_string(f"{lo_var} = {smem}[{lo_flat}]"))
-    cg.add_statement(statement_from_string(f"{hi_var} = {smem}[{hi_flat}]"))
-
-    return [
-        expr_from_string(lo_var),
-        expr_from_string(hi_var),
-    ]
-
-
 @_decorators.ref(split)
 def _(tensor: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
     return cast("tuple[torch.Tensor, torch.Tensor]", torch.unbind(tensor, dim=-1))
@@ -271,27 +187,6 @@ def _(tensor0: torch.Tensor, tensor1: torch.Tensor) -> torch.Tensor:
     return tensor0.new_empty([*broadcast_shape, 2])
 
 
-@_decorators.codegen(join, "cute")
-def _(state: CodegenState) -> ast.AST:
-    from .._compiler.cute.cute_reshape import _get_dim_local_coord
-    from .._compiler.generate_ast import GenerateAST
-
-    fx_node = state.fx_node
-    assert fx_node is not None
-    output_val = fx_node.meta["val"]
-    assert isinstance(output_val, torch.Tensor)
-    assert isinstance(state.codegen, GenerateAST)
-
-    new_dim = output_val.ndim - 1
-    selector = _get_dim_local_coord(state.codegen, output_val, new_dim)
-
-    return expr_from_string(
-        f"(({{a}}) if ({selector}) == cutlass.Int32(0) else ({{b}}))",
-        a=state.ast_arg(0),
-        b=state.ast_arg(1),
-    )
-
-
 @_decorators.ref(join)
 def _(tensor0: torch.Tensor, tensor1: torch.Tensor) -> torch.Tensor:
     left, right = torch.broadcast_tensors(tensor0, tensor1)
@@ -303,4 +198,5 @@ def _(tensor0: torch.Tensor, tensor1: torch.Tensor) -> torch.Tensor:
 # helion/_compiler/<backend>/.  Import them here (at module import time) so the
 # @_decorators.codegen(op, "<backend>") registrations run with the same eager
 # timing as when the bodies lived in this file -- no behavior change.
+import helion._compiler.cute.view_ops  # noqa: E402, F401
 import helion._compiler.triton.view_ops  # noqa: E402, F401
